+++ NEWS.md
@@ -1,5 +1,9 @@
 # tidystringdist 0.0.0.9000
 
+2017-10 
+
+- Col name in output is now the name of the method used. 
+
 * first commit